rust-backend/socketio_requirements.txt
python-socketio>=5.11.0
aiohttp>=3.9.0

